refactor crater code and bug fixes

+++ tileimage/RandomEx.cs
@@ -10,17 +10,17 @@ namespace tileimage
     public class RandomEx
     {
         RNGCryptoServiceProvider provider = new RNGCryptoServiceProvider();
-        public double NextDouble()
+        public float NextFloat()
         {
             var byteArray = new byte[4];
             provider.GetBytes(byteArray);
             UInt32 rand = BitConverter.ToUInt32(byteArray, 0);
-            return rand / (1.0 + UInt32.MaxValue);
+            return (float)( rand / (1.0 + UInt32.MaxValue));
         }
 
-        public double NextDouble(double min, double max)
+        public float NextFloat(float min, float max)
         {
-            return NextDouble() * (max - min) + min;
+            return NextFloat() * (max - min) + min;
         }
     }
 }
